// 05032026/sce-front/src/app/model/reportes/filtroReporteAuditoriaDigitacion.ts
export class FiltroReporteAuditoriaDigitacion {
  idProceso: number;
  idEleccion: string;
  idAmbito: number;
  ambitoElectoral: string;
  codigoAmbitoElectoral: string;
  idCentroComputo: number;
  codigoCentroComputo: string;
  centroComputo: string;
  departamento: string;
  provincia: string;
  idUbigeo: number;
  mesa: string;
  proceso?: string;
  eleccion?: string;
  schema?: string;
  distrito?:string;
  preferencial?: number;
  acronimo?: string;
  codigoEleccion: string;
}